Adicionado o cyclescontext
