+++ 02-ignite-timer/src/pages/History/index.tsx
@@ -1,6 +1,9 @@
+import { useContext } from 'react'
 import { HistoryContainer, HistoryList, Status } from './styles'
+import { CycleContext } from '../Home'
 
 export function History() {
+  const { cycles } = useContext(CycleContext)
   return (
     <HistoryContainer>
       <h1>Meu histórico</h1>
